requirements.txt
gradio
torch
numpy
pillow
albumentations
pandas
pytorch-lightning
timm
wandb